Project/dashboard/requirements.txt
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
google-cloud-bigquery==3.27.0
google-auth==2.37.0
pyarrow==18.1.0
db-dtypes==1.3.1
